avoid exporting ddm lib via client
this torpedos omicron builds because it pulls in dendrite
