+++ ddm/Cargo.toml
@@ -26,3 +26,4 @@ dendrite-common.workspace = true
 opte-ioctl.workspace = true
 oxide-vpc.workspace = true
 sled.workspace = true
+mg-common.workspace = true
